Refactor - Move all dot node stuff into dotnode package
