+++ src/main/scala/dotvisualizer/dotnodes/DotNode.scala
@@ -1,6 +1,4 @@
-// See LICENSE for license details.
-
-package dotvisualizer
+package dotvisualizer.dotnodes
 
 import scala.collection.mutable.ArrayBuffer
