+++ src/main/scala/dotvisualizer/dotnodes/ModuleNode.scala
@@ -1,6 +1,6 @@
 // See LICENSE for license details.
 
-package dotvisualizer
+package dotvisualizer.dotnodes
 
 import java.io.{File, PrintWriter}
 
@@ -60,7 +60,7 @@ case class ModuleNode(name: String, parentOpt: Option[DotNode], var url_string:O
   }
 }
 
-import sys.process._
+import scala.sys.process._
 
 //noinspection ScalaStyle
 object ModuleNode {
