+++ src/main/scala/dotvisualizer/dotnodes/NodeNode.scala
@@ -1,6 +1,6 @@
 // See LICENSE for license details.
 
-package dotvisualizer
+package dotvisualizer.dotnodes
 
 case class NodeNode(name: String, parentOpt: Option[DotNode]) extends DotNode {
   def render: String = s"""$absoluteName [label = "$name" shape="rectangle"]; """
