+++ src/main/scala/dotvisualizer/dotnodes/PrimOpNode.scala
@@ -1,6 +1,8 @@
 // See LICENSE for license details.
 
-package dotvisualizer
+package dotvisualizer.dotnodes
+
+import dotvisualizer.dotnodes.DotNode
 
 object PrimOpNode {
   /* the pseudoHash is necessary because case classes with identical args have identical hashes */
