+++ src/main/scala/dotvisualizer/dotnodes/RegisterNode.scala
@@ -1,6 +1,6 @@
 // See LICENSE for license details.
 
-package dotvisualizer
+package dotvisualizer.dotnodes
 
 case class RegisterNode(name: String, parentOpt: Option[DotNode]) extends DotNode {
   override val in: String = s"struct_$absoluteName:in"
